#3 delete List finished
