+++ src/Login/backend.ts
@@ -32,3 +32,15 @@ export async function getList() {
         .then((response) => response.json())
         .then((data) => data.shoppingLists)
 }
+
+export async function deleteList(id: number) {
+    const secret = urlCheck();
+    return await fetch(`http://localhost:3000/overview/${id}`, {
+        method: "DELETE",
+        headers: {
+            "Content-Type": "application/json",
+            "x-shared-shopper-secret": `${secret}`,
+            "authorization": `${session.token}`,
+        }
+    })
+}
